# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TucanoTime API",
    "version": "0.1.0",
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "description": "Timesheet system: daily/weekly time entries, customers and project codes with per-entity currencies and hourly rates. File-backed (no database); the API is the only writer below TUCANO_DATA_DIR. Access is private-per-user (#51): a member sees and edits only their own entries/expenses/submissions; an administrator sees all data and owns invoicing, approvals and user management (the /users and /invoices routes return 403 for members). Mutating endpoints may return 503 (Retry-After) under write contention (#62)."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tt_session",
        "description": "HttpOnly session cookie issued by /auth/login or /auth/bootstrap."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "validation_failed",
                  "not_found",
                  "conflict",
                  "bad_request",
                  "internal",
                  "invalid_json",
                  "write_lock_busy",
                  "lock_unavailable"
                ]
              },
              "message": {
                "type": "string",
                "description": "Human-readable; never contains paths, stack traces or raw filesystem errors."
              },
              "fields": {
                "type": "array",
                "description": "Per-field problems for validation_failed.",
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                }
              }
            }
          }
        }
      },
      "FieldError": {
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Currency": {
        "type": "string",
        "pattern": "^[A-Z]{3}$",
        "description": "ISO-4217 alpha-3 code, upper case.",
        "examples": [
          "EUR",
          "USD"
        ]
      },
      "ProjectCode": {
        "type": "string",
        "pattern": "^[A-Z0-9][A-Z0-9_-]{0,23}$",
        "description": "Normalised upper case; letters and digits, dashes and underscores after the first character; at most 24 characters.",
        "examples": [
          "ACME-42"
        ]
      },
      "Hours": {
        "type": "number",
        "minimum": 0.01,
        "maximum": 24,
        "description": "Hours as a decimal with at most two fractional digits (0.01\u201324.00). Stored as integer hundredths.",
        "examples": [
          7.5
        ]
      },
...
      "Entry": {
        "type": "object",
        "required": [
          "id",
          "date",
          "customer_id",
          "project_code",
          "hours",
          "note",
          "billable",
          "source",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          },
          "customer_id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "oneOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "description": "The user who logged the entry (attribution, #21)."
          },
          "project_code": {
            "$ref": "#/components/schemas/ProjectCode"
          },
          "task_code": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ProjectCode"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional task within the project (#38)."
          },
          "hours": {
            "$ref": "#/components/schemas/Hours"
          },
          "note": {
            "type": "string",
            "maxLength": 500
          },
          "billable": {
            "type": "boolean",
            "description": "Billable vs non-billable time. Non-billable time is never invoiced."
          },
          "source": {
            "type": "string",
            "enum": [
              "manual",
              "timer",
              "calendar-import"
            ],
            "description": "Provenance. Server-set; API clients create manual entries."
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "EntryInput": {
        "type": "object",
        "required": [
          "date",
          "customer_id",
          "project_code",
          "hours"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          },
          "customer_id": {
            "type": "string",
            "format": "uuid"
          },
          "project_code": {
            "$ref": "#/components/schemas/ProjectCode"
          },
          "task_code": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ProjectCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "hours": {
            "$ref": "#/components/schemas/Hours"
          },
          "note": {
            "type": "string",
            "maxLength": 500,
            "default": ""
          },
          "billable": {
            "type": "boolean",
            "default": true
          }
        },
        "additionalProperties": false
      },
      "SummaryRow": {
        "type": "object",
        "required": [
          "key",
          "label",
          "currency",
          "hours",
          "amount_minor",
          "entries"
        ],
        "properties": {
          "key": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          },
          "hours": {
            "type": "number"
          },
          "amount_minor": {
            "type": "integer",
            "minimum": 0
          },
          "entries": {
            "type": "integer",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
      "Summary": {
        "type": "object",
        "required": [
          "group",
          "rows",
          "total_hours",
          "billable_hours",
          "nonbillable_hours"
        ],
        "properties": {
          "group": {
            "type": "string",
            "enum": [
              "customer",
              "project",
              "person",
              "week"
            ]
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SummaryRow"
            }
          },
          "total_hours": {
            "type": "number"
          },
          "billable_hours": {
            "type": "number"
          },
          "nonbillable_hours": {
            "type": "number"
          }
        },
        "additionalProperties": false
      },
...
    },
    "responses": {
      "ValidationError": {
        "description": "Payload failed validation before any write or side effect.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "The resource does not exist.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "BadRequest": {
        "description": "Malformed query or body.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "parameters": {
      "DateParam": {
        "name": "date",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "format": "date",
          "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
        },
        "description": "Single day, YYYY-MM-DD. Exactly one of date or from+to."
      },
      "FromParam": {
        "name": "from",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "format": "date",
          "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
        },
        "description": "Inclusive start of a range. Requires to; at most 400 days."
      },
      "ToParam": {
        "name": "to",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "format": "date",
          "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
        },
        "description": "Inclusive end of a range."
      }
    }
  },
  "paths": {
...
    "/entries": {
      "parameters": [
        {
          "$ref": "#/components/parameters/DateParam"
        },
        {
          "$ref": "#/components/parameters/FromParam"
        },
        {
          "$ref": "#/components/parameters/ToParam"
        }
      ],
      "get": {
        "operationId": "listEntries",
        "tags": [
          "entries"
        ],
        "summary": "List entries for a day or an inclusive date range",
        "description": "Exactly one form: ?date=YYYY-MM-DD or ?from=...&to=... (span at most 400 days).",
        "responses": {
          "200": {
            "description": "Matching entries, ordered by date then creation.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "entries"
                  ],
                  "properties": {
                    "entries": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Entry"
                      }
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      },
      "post": {
        "operationId": "createEntry",
        "tags": [
          "entries"
        ],
        "summary": "Add a time entry",
        "description": "Cross-references are validated before any write: the customer must exist and the project must belong to it.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntryInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created; server assigns id and timestamps.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Entry"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/entries/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getEntry",
        "tags": [
          "entries"
        ],
        "summary": "Fetch one entry",
        "responses": {
          "200": {
            "description": "The entry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Entry"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "operationId": "updateEntry",
        "tags": [
          "entries"
        ],
        "summary": "Full replacement of an entry (may change its day)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntryInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Entry"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      },
      "delete": {
        "operationId": "deleteEntry",
        "tags": [
          "entries"
        ],
        "summary": "Delete an entry",
        "responses": {
          "204": {
            "description": "Deleted."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/reports/summary": {
      "parameters": [
        {
          "$ref": "#/components/parameters/FromParam"
        },
        {
          "$ref": "#/components/parameters/ToParam"
        },
        {
          "name": "group",
          "in": "query",
          "required": false,
          "schema": {
            "type": "string",
            "enum": [
              "customer",
              "project",
              "person",
              "week"
            ],
            "default": "customer"
          }
        },
        {
          "name": "billable",
          "in": "query",
          "required": false,
          "schema": {
            "type": "boolean"
          },
          "description": "Scope rows to billable (true) or non-billable (false); omit for all."
        }
      ],
      "get": {
        "operationId": "summary",
        "tags": [
          "reports"
        ],
        "summary": "Hours and amounts grouped by customer, project or ISO week",
        "description": "Amounts are computed with each entry's effective rate (project override, else customer default) at read time.",
        "responses": {
          "200": {
            "description": "Aggregated rows.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Summary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/reports/export.csv": {
      "parameters": [
        {
          "$ref": "#/components/parameters/FromParam"
        },
        {
          "$ref": "#/components/parameters/ToParam"
        },
        {
          "name": "customer_id",
          "in": "query",
          "required": false,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "exportCsv",
        "tags": [
          "reports"
        ],
        "summary": "One CSV row per entry: date, customer, code, hours, currency, hourly_rate, amount, note",
        "responses": {
          "200": {
            "description": "CSV document (formula-injection escaped).",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
    "/reports/profitability": {
      "parameters": [
        {
          "$ref": "#/components/parameters/FromParam"
        },
        {
          "$ref": "#/components/parameters/ToParam"
        }
      ],
      "get": {
        "operationId": "profitability",
        "tags": [
          "reports"
        ],
        "summary": "Revenue vs cost (expenses + labour at cost rates) per customer (#29)",
        "responses": {
          "200": {
            "description": "Profitability rows.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "rows"
                  ],
                  "properties": {
                    "rows": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "customer_id",
                          "label",
                          "currency",
                          "revenue_minor",
                          "cost_minor",
                          "margin_minor"
                        ],
                        "properties": {
                          "customer_id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "label": {
                            "type": "string"
                          },
                          "currency": {
                            "type": "string"
                          },
                          "revenue_minor": {
                            "type": "integer"
                          },
                          "cost_minor": {
                            "type": "integer"
                          },
                          "margin_minor": {
                            "type": "integer"
                          }
                        },
                        "additionalProperties": false
                      }
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
  }
}